+++ apiphany-core/src/main/java/org/apiphany/meters/micrometer/MicrometerLibrary.java
@@ -1,6 +1,6 @@
 package org.apiphany.meters.micrometer;
 
-import org.apiphany.lang.Pair;
+import org.apiphany.lang.LibraryDescriptor;
 import org.apiphany.meters.MeterFactory;
 import org.morphix.reflection.Constructors;
 import org.morphix.reflection.Reflection;
@@ -21,13 +21,15 @@ public class MicrometerLibrary {
 	private static final String MICROMETER_METER_CLASS_NAME = "io.micrometer.core.instrument.Meter";
 
 	/**
-	 * A {@link Pair} indicating whether the Micrometer library is present on the classpath, along with the
+	 * A {@link LibraryDescriptor} indicating whether the Micrometer library is present on the classpath, along with the
 	 * {@link MeterFactory} implementation class to use if it is available.
 	 * <p>
-	 * The {@code Boolean} value is {@code true} if Micrometer is detected, {@code false} otherwise.
+	 * The {@link LibraryDescriptor#isPresent()} value is {@code true} if Micrometer is detected, {@code false} otherwise.
 	 */
-	public static final Pair<Boolean, Class<? extends MeterFactory>> INFORMATION =
-			Pair.of(Reflection.isClassPresent(MICROMETER_METER_CLASS_NAME), MicrometerFactory.class);
+	public static final LibraryDescriptor<? extends MeterFactory> DESCRIPTOR =
+			LibraryDescriptor.of(
+					Reflection.isClassPresent(MICROMETER_METER_CLASS_NAME),
+					MicrometerFactory.class);
 
 	/**
 	 * Private constructor to prevent instantiation.